+++ source/stack.inl
@@ -28,15 +28,15 @@ void Stack<T>::Free(F&& freeElement) {
 // ***********************************************************************
 
 template<typename T>
-void Stack<T>::Resize(uint32_t desiredCount) {
-    array.Resize(desiredCount);
+void Stack<T>::Resize(u64 desiredCount) {
+    array.Resize((u32)desiredCount);
 }
 
 // ***********************************************************************
 
 template<typename T>
-void Stack<T>::Reserve(uint32_t desiredCapacity) {
-    array.Reserve(desiredCapacity);
+void Stack<T>::Reserve(u64 desiredCapacity) {
+    array.Reserve((u32)desiredCapacity);
 }
 
 // ***********************************************************************
@@ -65,11 +65,11 @@ T& Stack<T>::Top() {
 // ***********************************************************************
 
 template<typename T>
-T& Stack<T>::operator[](int32_t i) {
+T& Stack<T>::operator[](i64 i) {
     if (i >= 0) {
         return array[i];
     } else {
-        Assert(i >= -int32_t(array.count));
+        Assert(i >= -i64(array.count));
         return array[array.count + i];
     }
 }
@@ -77,11 +77,11 @@ T& Stack<T>::operator[](int32_t i) {
 // ***********************************************************************
 
 template<typename T>
-const T& Stack<T>::operator[](int32_t i) const {
+const T& Stack<T>::operator[](i64 i) const {
     if (i >= 0) {
         return array[i];
     } else {
-        Assert(i >= -int32_t(array.count));
+        Assert(i >= -i64(array.count));
         return array[array.count + i];
     }
 }
